added train, val test split in preprocessing

+++ dvc_workshop/pipeline/preprocess/io.py
@@ -3,7 +3,12 @@ import shutil
 import glob
 from typing import List, Iterable
 import numpy.typing as npt
-
+import numpy as np
+import ast
+import pandas as pd
+from git import Repo
+from PIL import Image
+from dvc_workshop.pipeline.preprocess.utils import perform_stratification
 
 
 def read_images(source_directory: str) -> List[str] : 
@@ -31,6 +36,28 @@ def save_images(images:Iterable[str],target_directory:str) :
         os.makedirs(target_directory)
     # copy selected images
     for img in images :
-        shutil.copy(img,target_directory)
+        if not os.path.isfile(os.path.join(target_directory,img)):
+            shutil.copy(img,target_directory)
+
+
+def generate_dataset(source_file: str,images_directory:str,target_directory:str) -> None:
+    """Void function to generate stratified csv f
+    ile with train test and validation data"""
+    path_labels_df = pd.read_csv(
+        source_file+ "/train.csv",
+        usecols=["Id", "Genre"],
+        converters={"Genre": ast.literal_eval},
+    )
+    path_labels_df = path_labels_df[["Id", "Genre"]]
+    path_labels_df.columns = ["Paths", "Labels"]
+    path_labels_df["Paths"] = (
+       images_directory+ path_labels_df["Paths"] + ".jpg"
+    )
+
+    train, test, valid = perform_stratification(path_labels_df, 0.3, 0.8, 1)
+    train.to_csv(os.path.join(target_directory ,"train.csv"), index=False)
+    test.to_csv(os.path.join(target_directory ,"test.csv"), index=False)
+    valid.to_csv(os.path.join(target_directory ,"valid.csv"), index=False)
+
